requirements.txt
Django>=4.0,<5.0
Pillow>=9.0
django-crispy-forms>=1.14
crispy-bootstrap4>=2022.1
